refactor: extract RubricSelector from AssistantForm (SRP)
